Force v4/v6 server lookup
This commit adds support for forcing whois to use v4 or v6 when looking up the
whois-server. There is no special handling of the case where a server is
specified manually. I assume a user to be smart enough to know what he or she is
doing when the flags are used.

One use-case for this feature is as follows:

I am working with a testbed where we frequently use traceroute to fetch the
trace to different servers, and then whois is used to lookup unknown IP
addresses. The nodes in the testbed are connected to a multitude of networks,
some v4-only, some v6-only and most v4/v6.

Several of the v6 networks are, for some reason hidden inside the ISPs network,
unstable. v4, on the other hand, is stable accross all networks. Linux, when
getaddrinfo is used with AF_UNSPEC, prioritizes v6 when a domain name is
resolved. When connected to an unstable v6-network, whois takes at least 10
second (the default connect timeout) longer to finish than when forcing v4.