Update TLD servers: gi, lc, sc, vc
Apparently these registries switched at some point in the past to a thick
model, so there are no referrals anymore.
Due to this the program should not use the generic Afilias queries code,
because the missing referral to the registry whois server would make it
return an error status.

Currently the old server whois.afilias-grs.info and the new server
whois2.afilias-grs.net, which is listed in the IANA DB for all the TLDs
but .lc, have the same IP address.