Fix IP address query bug introduced by recovered list
ip_del.h uses network and bitmask fields to determine server
ip_del_recovered.h uses start ip and end ip

These were being joined in a single array and the code to
determine the server was the same, even though the data type
was different.

This patch separates out the recovered list and adds the
correct checks to determine a match.