Support CIDR queries to whois.arin.net
This is a behaviour change: in this case the client used to use the
parameters for non-CIDR queries, so the server would return nothing.
Now an exact match for the provided ip/length query is requested.

If the query is not an exact match then this is also different from the
default behaviour of the server, which is to return a non-verbose
summary of all less specific networks.