Update some TLD servers which reference Afilias
Afilias was bought by Donuts in 2020 and the combined entity is
now called Identity Digital. I went through each entry that has
'afilias' in it to see if it needs updating. This covers ones
that look straightforward: a sensible server name is in the IANA
whois info, and it points to Identity Digital's catch-all server
currently at 52.37.99.5 and 2600:1f13:101:c200:6a3:4cbe:e213:77ab

There are several more complicated cases that I left alone and
will file tickets for.

Signed-off-by: Adam Williamson <awilliam@redhat.com>